+++ src/pages/programRuleVariables/fields/SourceTypeField.tsx
@@ -2,45 +2,56 @@ import i18n from '@dhis2/d2-i18n'
 import { SingleSelectFieldFF } from '@dhis2/ui'
 import React from 'react'
 import { Field as FieldRFF } from 'react-final-form'
-import { PROGRAM_RULE_VARIABLE_SOURCE_TYPE } from '../../../lib'
+import { getConstantTranslation } from '../../../lib'
 import { ProgramRuleVariable } from '../../../types/generated'
 
-const {
-    DATAELEMENT_NEWEST_EVENT_PROGRAM_STAGE,
-    DATAELEMENT_NEWEST_EVENT_PROGRAM,
-    DATAELEMENT_CURRENT_EVENT,
-    DATAELEMENT_PREVIOUS_EVENT,
-    CALCULATED_VALUE,
-    TEI_ATTRIBUTE,
-} = ProgramRuleVariable.programRuleVariableSourceType
-
-const SOURCE_TYPE_OPTIONS = [
+const SOURCE_TYPES = [
     {
-        value: DATAELEMENT_NEWEST_EVENT_PROGRAM_STAGE,
-        label: PROGRAM_RULE_VARIABLE_SOURCE_TYPE.DATAELEMENT_NEWEST_EVENT_PROGRAM_STAGE,
+        value: ProgramRuleVariable.programRuleVariableSourceType
+            .DATAELEMENT_NEWEST_EVENT_PROGRAM_STAGE,
+        label: getConstantTranslation(
+            ProgramRuleVariable.programRuleVariableSourceType
+                .DATAELEMENT_NEWEST_EVENT_PROGRAM_STAGE
+        ),
     },
     {
-        value: DATAELEMENT_NEWEST_EVENT_PROGRAM,
-        label: PROGRAM_RULE_VARIABLE_SOURCE_TYPE.DATAELEMENT_NEWEST_EVENT_PROGRAM,
+        value: ProgramRuleVariable.programRuleVariableSourceType
+            .DATAELEMENT_NEWEST_EVENT_PROGRAM,
+        label: getConstantTranslation(
+            ProgramRuleVariable.programRuleVariableSourceType
+                .DATAELEMENT_NEWEST_EVENT_PROGRAM
+        ),
     },
     {
-        value: DATAELEMENT_CURRENT_EVENT,
-        label: PROGRAM_RULE_VARIABLE_SOURCE_TYPE.DATAELEMENT_CURRENT_EVENT,
+        value: ProgramRuleVariable.programRuleVariableSourceType
+            .DATAELEMENT_CURRENT_EVENT,
+        label: getConstantTranslation(
+            ProgramRuleVariable.programRuleVariableSourceType
+                .DATAELEMENT_CURRENT_EVENT
+        ),
     },
     {
-        value: DATAELEMENT_PREVIOUS_EVENT,
-        label: PROGRAM_RULE_VARIABLE_SOURCE_TYPE.DATAELEMENT_PREVIOUS_EVENT,
+        value: ProgramRuleVariable.programRuleVariableSourceType
+            .DATAELEMENT_PREVIOUS_EVENT,
+        label: getConstantTranslation(
+            ProgramRuleVariable.programRuleVariableSourceType
+                .DATAELEMENT_PREVIOUS_EVENT
+        ),
     },
     {
-        value: CALCULATED_VALUE,
-        label: PROGRAM_RULE_VARIABLE_SOURCE_TYPE.CALCULATED_VALUE,
+        value: ProgramRuleVariable.programRuleVariableSourceType
+            .CALCULATED_VALUE,
+        label: getConstantTranslation(
+            ProgramRuleVariable.programRuleVariableSourceType.CALCULATED_VALUE
+        ),
     },
     {
-        value: TEI_ATTRIBUTE,
-        label: PROGRAM_RULE_VARIABLE_SOURCE_TYPE.TEI_ATTRIBUTE,
+        value: ProgramRuleVariable.programRuleVariableSourceType.TEI_ATTRIBUTE,
+        label: getConstantTranslation(
+            ProgramRuleVariable.programRuleVariableSourceType.TEI_ATTRIBUTE
+        ),
     },
 ]
-
 export function SourceTypeField({
     onChange,
 }: Readonly<{ onChange?: () => void }>) {
@@ -59,7 +70,7 @@ export function SourceTypeField({
                     meta={meta}
                     inputWidth="400px"
                     label={i18n.t('Source type')}
-                    options={SOURCE_TYPE_OPTIONS}
+                    options={SOURCE_TYPES}
                     placeholder={i18n.t('Select source type')}
                 />
             )}